m2mbase: change lwm2m_parameters_s::instance_id to 16 bit uint
There is no need for 32bit variable, just use the 16bit one.

+++ mbed-client/m2mbase.h
@@ -122,7 +122,7 @@ public:
     typedef struct lwm2m_parameters {
         //add multiple_instances
         uint32_t            max_age;
-        uint32_t            instance_id;
+        uint16_t            instance_id; // XXX: this is not properly aligned now, need to reorder these after the elimination is done
         int32_t             name_id;
         char*               name; //for backwards compability
         sn_nsdl_dynamic_resource_parameters_s *dynamic_resource_params;
